+++ src/pint/scripts/compare_parfiles.py
@@ -75,13 +75,22 @@ def main(argv=None):
     parser.add_argument(
         "--log-level",
         type=str,
-        choices=("TRACE", "DEBUG", "INFO", "WARNING", "ERROR"),
-        default="WARNING",
+        choices=pint.logging.levels,
+        default=pint.logging.script_level,
         help="Logging level",
         dest="loglevel",
     )
+    parser.add_argument(
+        "-v", "--verbosity", default=0, action="count", help="Increase output verbosity"
+    )
+    parser.add_argument(
+        "-q", "--quiet", default=0, action="count", help="Decrease output verbosity"
+    )
+
     args = parser.parse_args(argv)
-    pint.logging.setup(level=args.loglevel)
+    pint.logging.setup(
+        level=pint.logging.get_level(args.loglevel, args.verbosity, args.quiet)
+    )
 
     m1 = get_model(args.input1)
     m2 = get_model(args.input2)